# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sefest Api Workshop",
        "description": "api ini hanya diperuntukan untuk peserta dan pemateri workshop sefest 2025",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/order": {
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "create order",
                "operationId": "43c9166756ae9d15c72f801c0c5ec7e2",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "creator_id": {
                                                "type": "number"
                                            },
                                            "product_id": {
                                                "type": "number"
                                            },
                                            "total": {
                                                "type": "number"
                                            },
                                            "proof": {
                                                "type": "binary"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "creator_id": 1,
                                    "product_id": 1,
                                    "total": 5000,
                                    "proof": "image.png"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Order Berhasil dibu"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/creator/order/approve/{order}": {
            "put": {
                "tags": [
                    "Order"
                ],
                "summary": "approve Order",
                "operationId": "Title",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "Parameter with example",
                        "required": true,
                        "schema": {
                            "type": "int"
                        },
                        "examples": {
                            "int": {
                                "summary": "an int value",
                                "value": "1"
                            }
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "is_paid": {
                                                "type": "number"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "is_paid": "0"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "status berhasil diubah"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}